+++ finance_friend.client/package.json
@@ -18,7 +18,8 @@
     "@reduxjs/toolkit": "^2.2.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-redux": "^9.1.2"
+    "react-redux": "^9.1.2",
+    "react-router-dom": "^6.23.1"
   },
   "devDependencies": {
     "@types/node": "^20.12.2",
